[SPARK-2450] [CORE] Adds exeuctor log links to Web UI
Adds links to stderr/stdout in the executor tab
of the webUI for:
1) Standalone
2) Yarn client
3) Yarn cluster

This tries to add the log url support in a general
way so as to make it easy to add support for all the
cluster managers. This is done by using environment
variables to pass to the executor the log urls. The
SPARK_LOG_URL_ prefix is used and so additional logs
besides stderr/stdout can also be added.

To propagate this information to the UI we use
the onExecutorAdded spark listener event.

Although this commit doesn't add log urls when running
on a mesos cluster, it should be possible to add using
the same mechanism.